# Conflicts:
#	Belireview/target/m2e-wtp/web-resources/META-INF/maven/believe.review/brw/pom.properties

+++ Belireview/src/main/java/believe/review/brw/common/util/FileUtils.java
@@ -82,25 +82,27 @@ public class FileUtils {
 			main_image = (String)map.get("MOVIE_MAIN_IMAGE");
 			content_image = ((String) map.get("MOVIE_GALLERY_IMAGE")).split(",");
 		}else if(media.equals("ad")) {
-			poster_image = (String)map.get("AD_POSTER_IMAGE");
 			main_image = (String)map.get("AD_MAIN_IMAGE");
-			content_image = ((String) map.get("AD_CONTENT_IMAGE")).split(",");
 		}
 		
-		file = new File(filePath + "poster\\" + poster_image);
-		file.delete();
-		file = new File(filePath + "main\\" + main_image);
-		file.delete();
+		if(!media.equals("ad")) {
+			file = new File(filePath + "poster\\" + poster_image);
+			file.delete();
+			file = new File(filePath + "main\\" + main_image);
+			file.delete();
 		
-		for(int i = 0 ; i < content_image.length ; i++) {
-			file = new File(filePath + "content\\" + content_image[i]);
+			for(int i = 0 ; i < content_image.length ; i++) {
+				file = new File(filePath + "content\\" + content_image[i]);
+				file.delete();
+			}
+		}else {
+			file = new File(filePath + "main\\" + main_image);
 			file.delete();
 		}
 	}
 
 	//user용
-	//private static final String filePath_P = "C:\\인영\\sts\\Belireview\\Belireview\\src\\main\\webapp\\resources\\images\\user_profile\\";
-	private static final String filePath_P = "C:\\Users\\박재연\\Desktop\\Belireview\\Belireview\\src\\main\\webapp\\resources\\images\\user_profile\\";
+	private static final String filePath_P = "C:\\인영\\sts\\Belireview\\Belireview\\src\\main\\webapp\\resources\\images\\user_profile\\";
 	
 	public Map<String, Object> parseInsertFileInfo(Map<String, Object> map, HttpServletRequest request)
 			throws Exception {
